test/inspector/debugger/wasm-step-a-lot-expected.txt

Tests repeated stepping through a large function (should not OOM)

Running test: test
Setting up global instance variable.
Got wasm script: wasm://wasm/46f40116
Setting breakpoint
Paused 500 and running...
Paused 1000 and running...
Paused 1500 and running...
Paused 2000 and running...
Paused 2500 and running...
Paused 3000 and running...
Paused 3500 and running...
Paused 4000 and running...
test function returned.
Paused 4003 times.
